1>main.obj : error LNK2019: unresolved external symbol "public: __thiscall calc_D_matrix::~calc_D_matrix(void)" (??1calc_D_matrix@@QAE@XZ) r

+++ Adam's Project/calc_D_Matrix.cpp
@@ -2,7 +2,7 @@
 #include "calc_D_matrix.h"
 #include "Controller.h"
 
-calc_D_matrix::calc_D_matrix(int val1, float val2, float val3, Controller &controller)
+calc_D_matrix::calc_D_matrix(int val1, double val2, double val3, Controller &controller)
 {
 	ProblemType = val1;
 	YM = val2;
