+++ Adam's Project/calc_D_matrix.h
@@ -16,7 +16,7 @@ protected:
 	double Poisson;
 
 public:
-	calc_D_matrix(int val1, float val2, float val3, Controller &controller);
+	calc_D_matrix(int val1, double val2, double val3, Controller &controller);
 	~calc_D_matrix();
 	
 	//Functions
